requirements.txt
requests
python-dotenv
langchain_google_genai
google-generativeai